Made the verification fragment an activity
